feat: 去除docNO唯一限制；主键用long-bigint(20)；前后端用fastjson转化string-long

+++ src/main/java/com/internetplus/bankpunishment/bl/BankPunishmentBl.java
@@ -12,17 +12,17 @@ import java.util.List;
  * @date: 2021-10-20
  **/
 public interface BankPunishmentBl {
-    Integer insertBankPunishment(BankPunishment bankPunishment);
+    Long insertBankPunishment(BankPunishment bankPunishment);
 
     void updateBankPunishment(BankPunishment bankPunishment);
 
     void updateBankPunishmentExceptNull(BankPunishment bankPunishment);
 
-    void deleteBankPunishment(Integer id);
+    void deleteBankPunishment(Long id);
 
-    void publishBankPunishment(Integer id);
+    void publishBankPunishment(Long id);
 
-    BankPunishment selectBankPunishmentById(Integer id);
+    BankPunishment selectBankPunishmentById(Long id);
 
     List<BankPunishment> selectBankPunishment(BankPunishment bankPunishment);
 }
